Allow and configure hiding the intro on mobile pages
The intro section can be quite long, especially when a picture is
enabled. By default the intro appears on the left on computer screen
sizes, but above the content on mobile devices. This commit adds
configuration parameters, css, and template logic to hide the intro on
mobile devices, and configure if it is always hidden, or just on
content pages that are not the homepage.