Promote list layouts from post to default
Since the theme is list centric, other types of content will need list
entries to be rendered correctly. The _default/list.html template
refers to the content-list template, which previously was only defined
for posts. Any new content types a user adds would not show up in
lists (i.e. on the main page!) which is especially confusing for
people new to hugo.

This commit promotes the `content-list.html` template up from posts to
the _defaults category, along with the `featured.html` and
`stats.html` templates that it refers to. This allows other content
types (like blog) which are in the mainSections config parameter to
render their previews correctly.