Rename "on mobile" to "single column"
When referring to intro hiding, the intended behavior is to hide the
intro on all single column layouts, not just ones "on mobile." Notably
the `for-laptup-up` media type used the single column layout, but the
name and intention is for "non-mobile" devices. To more clearly
communicate intent this commit renames hideOnMobile to
hideWhenSingleColumn and changes the relevant references from mobile
hiding to single column hiding.