Separate homepage template from section lists
The homepage should display all content types in mainSections, but
individual sections should probably only display the pages
corresponding to their section. This corresponds to the use case and
confusion reported by @giom-mallet in #119.

This commit creates a separate homepage layout and modifies the list
layout to only display pages in a particular section. The homepage and
the list template share a lot of duplicated code percentage wise, but
it's only 20 lines so I was not sure if it would be worth it to add
another layer of templating and abstraction to remove this duplication
since most of the list template is already broken into partials.