Allow and configure hiding the intro on mobile pages (#108)
* Allow and configure hiding the intro on mobile pages

The intro section can be quite long, especially when a picture is
enabled. By default the intro appears on the left on computer screen
sizes, but above the content on mobile devices. This commit adds
configuration parameters, css, and template logic to hide the intro on
mobile devices, and configure if it is always hidden, or just on
content pages that are not the homepage.

* Disable intro hiding by default

This commit sets the example site into hiding parameters to False to
better track the exisiting theme behavior for backwards
compatibility. If these parameters are not included in a site config,
the intro will also not be hidden, maintaining existing behavior.

* Rename "on mobile" to "single column"

When referring to intro hiding, the intended behavior is to hide the
intro on all single column layouts, not just ones "on mobile." Notably
the `for-laptup-up` media type used the single column layout, but the
name and intention is for "non-mobile" devices. To more clearly
communicate intent this commit renames hideOnMobile to
hideWhenSingleColumn and changes the relevant references from mobile
hiding to single column hiding.