Address edge cases around multiple content types (#122)
* Update taxonomy to support multiple content types

Hugo allows multiple content types to specify tags and
categories. List all pages in the defined "mainSections" variable that
match a given category or tag in the appropriate taxonomies.

show all types of tagged content in the taxonomy

* Promote list layouts from post to default

Since the theme is list centric, other types of content will need list
entries to be rendered correctly. The _default/list.html template
refers to the content-list template, which previously was only defined
for posts. Any new content types a user adds would not show up in
lists (i.e. on the main page!) which is especially confusing for
people new to hugo.

This commit promotes the `content-list.html` template up from posts to
the _defaults category, along with the `featured.html` and
`stats.html` templates that it refers to. This allows other content
types (like blog) which are in the mainSections config parameter to
render their previews correctly.

* Separate homepage template from section lists

The homepage should display all content types in mainSections, but
individual sections should probably only display the pages
corresponding to their section. This corresponds to the use case and
confusion reported by @giom-mallet in #119.

This commit creates a separate homepage layout and modifies the list
layout to only display pages in a particular section. The homepage and
the list template share a lot of duplicated code percentage wise, but
it's only 20 lines so I was not sure if it would be worth it to add
another layer of templating and abstraction to remove this duplication
since most of the list template is already broken into partials.

* Make default layouts fully featured

Previously the default page layouts were not fully featured like the
post type. Notably, they lacked comments, pagination support, social
media sharing buttons, and a featured image. This commit makes default
content types include all of these features that were previously
restricted to posts only. It achieves this goal by moving most post
functionality to the _default content type, allowing post to just be
another _default.

* Remove explicit page types from examples and archetypes

With post formatting now the default type, we no longer need to
expicitly specify type=post in the front matter of the content files
even for non-post sections of the site.

* Clarify the role of the mainSections parameter

This commit adds a clarifying comment to the example config.toml and
removes the non-existant "post" section from the mainSections
parameter for clarity.